Reverted commit fa28e4dc91ffefb79a4601408438ff1ab96f620e

+++ phoenix-core/src/test/java/org/apache/phoenix/query/BaseTest.java
@@ -86,7 +86,6 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
-import java.io.IOException;
 import java.math.BigDecimal;
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
@@ -109,20 +108,20 @@ import java.util.Set;
 
 import javax.annotation.Nonnull;
 
-import org.apache.commons.lang.SystemUtils;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.HBaseConfiguration;
+import org.apache.hadoop.hbase.HBaseTestingUtility;
 import org.apache.hadoop.hbase.HColumnDescriptor;
 import org.apache.hadoop.hbase.HConstants;
 import org.apache.hadoop.hbase.HTableDescriptor;
 import org.apache.hadoop.hbase.IntegrationTestingUtility;
-import org.apache.hadoop.hbase.TableNotEnabledException;
 import org.apache.hadoop.hbase.client.HBaseAdmin;
 import org.apache.hadoop.hbase.coprocessor.RegionServerObserver;
 import org.apache.hadoop.hbase.io.encoding.DataBlockEncoding;
 import org.apache.hadoop.hbase.ipc.PhoenixRpcSchedulerFactory;
 import org.apache.hadoop.hbase.ipc.RpcControllerFactory;
 import org.apache.hadoop.hbase.ipc.controller.ServerRpcControllerFactory;
+import org.apache.hadoop.hbase.regionserver.HRegionServer;
 import org.apache.hadoop.hbase.regionserver.LocalIndexMerger;
 import org.apache.hadoop.hbase.regionserver.RSRpcServices;
 import org.apache.hadoop.hbase.util.Bytes;
@@ -166,10 +165,6 @@ import com.google.common.collect.Sets;
 public abstract class BaseTest {
     private static final Map<String,String> tableDDLMap;
     private static final Logger logger = LoggerFactory.getLogger(BaseTest.class);
-
-    // TIME_CLOCK_SKEW_THRESHOLD is used in tests to compensate tests where local time are used to
-    // compare time from another server
-    public static long TIME_CLOCK_SKEW_THRESHOLD = 3000;
 
     static {
         ImmutableMap.Builder<String,String> builder = ImmutableMap.builder();
@@ -478,7 +473,7 @@ public abstract class BaseTest {
     private static String url;
     protected static PhoenixTestDriver driver;
     private static boolean clusterInitialized = false;
-    private static IntegrationTestingUtility utility;
+    private static HBaseTestingUtility utility;
     protected static final Configuration config = HBaseConfiguration.create(); 
     
     protected static String getUrl() {
@@ -519,13 +514,11 @@ public abstract class BaseTest {
         }
     }
     
-    protected static void dropAllTables() throws Exception {
-        HBaseAdmin admin = driver.getConnectionQueryServices(null, null).getAdmin();
+    protected static void dropNonSystemTables() throws Exception {
         try {
-            destroyDriver();
+            disableAndDropNonSystemTables();
         } finally {
-            driver = null;
-            disableAndDropTables(admin);
+            destroyDriver();
         }
     }
 
@@ -573,7 +566,7 @@ public abstract class BaseTest {
      */
     private static String initMiniCluster(Configuration conf, ReadOnlyProps overrideProps) {
         setUpConfigForMiniCluster(conf, overrideProps);
-        utility = new IntegrationTestingUtility(conf);
+        utility = new HBaseTestingUtility(conf);
         try {
             utility.startMiniCluster(NUM_SLAVES_BASE);
             // add shutdown hook to kill the mini cluster
@@ -593,7 +586,7 @@ public abstract class BaseTest {
         }
     }
 
-    protected static String getLocalClusterUrl(IntegrationTestingUtility util) throws Exception {
+    protected static String getLocalClusterUrl(HBaseTestingUtility util) throws Exception {
         String url = QueryUtil.getConnectionUrl(new Properties(), util.getConfiguration());
         return url + PHOENIX_TEST_DRIVER_URL_PARAM;
     }
@@ -788,13 +781,8 @@ public abstract class BaseTest {
     }
     
     protected static void deletePriorTables(long ts, String tenantId, String url) throws Exception {
-        if(SystemUtils.IS_OS_WINDOWS) {
-            // deal with the windows low time resolution
-            Thread.sleep(200);
-        }
         Properties props = new Properties();
-        props.put(QueryServices.THREAD_POOL_SIZE_ATTRIB, Integer.toString(24));
-        props.put(QueryServices.QUEUE_SIZE_ATTRIB, Integer.toString(2048));
+        props.put(QueryServices.QUEUE_SIZE_ATTRIB, Integer.toString(1024));
         if (ts != HConstants.LATEST_TIMESTAMP) {
             props.setProperty(CURRENT_SCN_ATTRIB, Long.toString(ts));
         }
@@ -805,10 +793,6 @@ public abstract class BaseTest {
         }
         finally {
             conn.close();
-        }
-        if(SystemUtils.IS_OS_WINDOWS) {
-            // deal with the windows low time resolution
-            Thread.sleep(200);
         }
     }
     
@@ -845,7 +829,6 @@ public abstract class BaseTest {
                     logger.info("Table " + fullTableName + " is already deleted.");
                 }
             }
-            rs.close();
             if (lastTenantId != null) {
                 conn.close();
             }
@@ -877,7 +860,6 @@ public abstract class BaseTest {
             logger.info("DROP SEQUENCE STATEMENT: DROP SEQUENCE " + SchemaUtil.getEscapedTableName(rs.getString(2), rs.getString(3)));
             conn.createStatement().execute("DROP SEQUENCE " + SchemaUtil.getEscapedTableName(rs.getString(2), rs.getString(3)));
         }
-        rs.close();
     }
     
     protected static void initSumDoubleValues(byte[][] splits, String url) throws Exception {
@@ -1640,28 +1622,24 @@ public abstract class BaseTest {
         }
     }
     
-
     /**
-     * Disable and drop all the tables
+     * Disable and drop all the tables except SYSTEM.CATALOG and SYSTEM.SEQUENCE
      */
-    protected static void disableAndDropTables(HBaseAdmin admin) throws Exception {
+    private static void disableAndDropNonSystemTables() throws Exception {
+        HBaseAdmin admin = driver.getConnectionQueryServices(null, null).getAdmin();
         try {
             HTableDescriptor[] tables = admin.listTables();
             for (HTableDescriptor table : tables) {
-                try{
+                String schemaName = SchemaUtil.getSchemaNameFromFullName(table.getName());
+                if (!QueryConstants.SYSTEM_SCHEMA_NAME.equals(schemaName)) {
                     admin.disableTable(table.getName());
-                } catch(IOException ex) {
-                    if(ex instanceof TableNotEnabledException) {
-                        //ignored
-                    }
+                    admin.deleteTable(table.getName());
                 }
-                admin.deleteTable(table.getName());
             }
         } finally {
             admin.close();
         }
     }
-
 
     public static void assertOneOfValuesEqualsResultSet(ResultSet rs, List<List<Object>>... expectedResultsArray) throws SQLException {
         List<List<Object>> results = Lists.newArrayList();
@@ -1711,7 +1689,7 @@ public abstract class BaseTest {
         assertEquals(expectedCount, count);
     }
     
-    public IntegrationTestingUtility getUtility() {
+    public HBaseTestingUtility getUtility() {
         return utility;
     }
